- add board columns endpoints crud + reordering - add column permissions - change board column position to use numeric for more robust reordering logic

+++ backend/src/main/resources/db/migration/V5__create_board_columns.sql
@@ -2,7 +2,7 @@ CREATE TABLE board_columns(
     id UUID PRIMARY KEY,
     board_id UUID NOT NULL,
     name VARCHAR(100) NOT NULL,
-    position INTEGER NOT NULL,
+    position NUMERIC(30, 10) NOT NULL,
 
     FOREIGN KEY (board_id) REFERENCES boards(id) ON DELETE CASCADE
 );
